cleaning up cart manager

+++ models/engine/cart_manager.py
@@ -8,6 +8,9 @@ from models.cart import Cart
 class Cart_manager:
     @staticmethod
     def create_Cart(user_id, product_id):
+        """
+        creating a new cart
+        """
         cart = Cart(user_id=user_id, product_id=product_id)
         storage.new(cart)
         storage.save()
@@ -15,26 +18,44 @@ class Cart_manager:
 
     @staticmethod
     def read_Cart(cart_id):
+        """
+        listing cart content
+        """
+
         return storage.new_get(Cart, cart_id)
 
     @staticmethod
     def add_product(cart_id, product_id):
+        """
+        add product to cart
+        """
         cart = storage.new_get(Cart, cart_id)
-        cart.product_id = product_id
-        storage.save()
-        return cart
+        if product_id not in cart.product_id:
+            cart.product_id.append(product_id)
+            storage.save()
+            return cart
+
     @staticmethod
-    def delete_product(cart_id):
+    def delete_product(cart_id, product_id):
+        """
+        delete product in cart
+        """
+
         product = storage.new_get(Cart, product_id)
-        storage.delete(product_id)
-        storage.save()
-        return cart
+        if product_id in Cart.product_id:
+            storage.delete(product_id)
+            storage.save()
+            return cart
+        
     @staticmethod
     def delete_cart(cart_id):
+        """
+        delete the cart
+        """
         cart = storage.new_get(Cart, cart_id)
         storage.delete(cart)
         storage.save()
-        return cart
+        return None
